requirements-dev.txt
Pygments==2.0.2
git+https://github.com/TamiaLab/PySkCode.git
coverage==3.7.1
Django==1.8.6
django-debug-toolbar==1.4
Pillow==2.9.0
psycopg2==2.6.1
pylibmc==1.5.0
pytz==2015.4
six==1.9.0
